Move Environment into dedicated module
