+++ hydra-node/hydra-node.cabal
@@ -67,6 +67,7 @@ library
     Hydra.Chain.Offline
     Hydra.ContestationPeriod
     Hydra.Crypto
+    Hydra.Environment
     Hydra.Events
     Hydra.HeadId
     Hydra.HeadLogic
